moved auth models into a new folder
