+++ client/src/util/authUtil.ts
@@ -1,7 +1,8 @@
 import { Dispatch, SetStateAction } from "react";
 import axios from "axios";
 import jwtDecode from "jwt-decode";
-import { UserProfileState, DecodedJWT } from "./models";
+import { UserProfileState } from "./models";
+import { DecodedJWT } from "./models/AuthModels";
 import { VERIFY_TOKEN_RESPONSE } from "./constants";
 
 export const setUserProfile = async (userJWT: string, setUserProfileState: Dispatch<SetStateAction<UserProfileState>>) => {
