+++ client/src/util/models/AuthModels.ts
@@ -1,0 +1,15 @@
+
+export interface DecodedJWT {
+  exp: number,
+  iat: number,
+  user: {
+    email: string,
+    _id: string
+  }
+}
+
+export interface UserProfileState {
+  userID: string;
+  userEmail: string;
+  token: string;
+}
